Move wrapper tests to correct folder, optionally bind functions
