+++ tests/shim/unit/all.ts
@@ -3,7 +3,7 @@ import '../../../src/shim/main';
 import './support/decorators';
 import './support/queue';
 import './support/util';
-import './support/wrapper';
+import './util/wrapper';
 import './AbortController';
 import './array';
 import './global';
